Create MoreEfficientMeritValCalc
This program is the latest meritValueCalculator version of my program as of April 9th. I found hardcoding pumpPowers and the experimentally derived data that came from it to be the most efficient way of calculating the merit value rather than porting the data and trying to repeatedly nest functions within each other. 

If you want to vary constants, you have to manually either create new functions (and swap out the variables in the variedVarName and endTime parts of the function).